LinkedListInsertion.py
here is THE REDO FROM SCRACH of a simple program ( LEARNED FROM THE INTERNET AND UNDERSTOOD ), its task is to create a linked list and insert values at the beginning of it (here the main first goal ws to be able to insert values at the front of the linked list and understand the mechanism of how it works), plus, we also know how to create a printing function in a class which helps us to print the linked list.

this will be relevant for the coming part of the lessons about data structures.