Update insertionReview.py
now we are inserting a new method called "insert_values" and this method will allow us to take a lit of values as a single input and it will create a new fresh linked list.  it will going to wipe up (erase them) all the current value from the list and input new values in that list.