Update insertionReview.py
here we are adding a for loop which will loop trough all the data or values in out inputted list. this loop comes to use our previous function (method we created called "insert_at_end").

and it is using that "insert_at_end" method coz it so that, when we want to input a list of values at the same time, each value from that inputted list will come after another until we reach the last value which will be added at the end (coz we used the insert_at_end method)